+++ Source/Airport/nbproject/private/private.xml
@@ -2,13 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/thebestp9/Desktop/Tourism-Project/Tourism-Project/Airport/src/airport/IndiaSOP.java</file>
-            <file>file:/C:/Users/thebestp9/Desktop/Tourism-Project/Tourism-Project/Airport/src/airport/Airport.java</file>
-            <file>file:/C:/Users/thebestp9/Desktop/Tourism-Project/Tourism-Project/Airport/src/airport/SopIndia.txt</file>
-            <file>file:/C:/Users/thebestp9/Desktop/Tourism-Project/Tourism-Project/Airport/src/airport/GeneralSOP.txt</file>
-            <file>file:/C:/Users/thebestp9/Desktop/Tourism-Project/Tourism-Project/Airport/src/airport/SopMalaysia.txt</file>
-            <file>file:/C:/Users/thebestp9/Desktop/Tourism-Project/Tourism-Project/Airport/src/airport/MalaySOP.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
