+++ Source/MainLauncher/nbproject/private/private.properties
@@ -5,6 +5,6 @@ compile.on.save=true
 <<<<<<< HEAD
 user.properties.file=C:\\Users\\user\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
 =======
-user.properties.file=C:\\Users\\thebestp9\\AppData\\Roaming\\NetBeans\\11.3\\build.properties
+user.properties.file=C:\\Users\\hp pc\\AppData\\Roaming\\NetBeans\\8.2rc\\build.properties
 >>>>>>> 33634fb4ee648b06946b304da605b206fac721d0
 >>>>>>> bb30a623774832618cc0de8bd029855f9f9895c1
